+++ src/main/java/com/cojac/storyteller/common/config/SecurityConfig.java
@@ -4,10 +4,10 @@ import com.cojac.storyteller.user.repository.LocalUserRepository;
 import com.cojac.storyteller.user.repository.SocialUserRepository;
 import com.cojac.storyteller.common.redis.RedisService;
 import com.cojac.storyteller.user.jwt.*;
-import com.cojac.storyteller.user.security.CustomAuthenticationEntryPoint;
-import com.cojac.storyteller.user.security.JWTFilter;
-import com.cojac.storyteller.user.security.LoginFilter;
-import com.cojac.storyteller.user.security.LogoutFilter;
+import com.cojac.storyteller.user.controller.security.CustomAuthenticationEntryPoint;
+import com.cojac.storyteller.user.jwt.JWTFilter;
+import com.cojac.storyteller.user.service.security.LoginFilter;
+import com.cojac.storyteller.user.service.security.LogoutFilter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
