+++ src/main/java/com/cojac/storyteller/user/controller/security/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.cojac.storyteller.user.security;
+package com.cojac.storyteller.user.controller.security;
 
 import com.cojac.storyteller.response.code.ErrorCode;
 import com.cojac.storyteller.response.dto.ErrorResponseDTO;
