+++ src/main/java/com/cojac/storyteller/user/jwt/JWTFilter.java
@@ -1,10 +1,9 @@
-package com.cojac.storyteller.user.security;
+package com.cojac.storyteller.user.jwt;
 
 import com.cojac.storyteller.response.code.ErrorCode;
 import com.cojac.storyteller.user.entity.LocalUserEntity;
 import com.cojac.storyteller.user.dto.security.CustomUserDetails;
 import com.cojac.storyteller.common.util.ErrorResponseUtil;
-import com.cojac.storyteller.user.jwt.JWTUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
